+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.1.0
 
 * Add in stacktrace-js for source maps in stack traces and hide Opal/Opal-RSpec/Karma code from test failure traces
+* Examples filtered out using `filter_run_including` now show up as skipped in Karma results
 * Remove undocumented source map results feature
 
 ## 1.0.10 (27 January 2016)